style: shorten docstrings to satisfy ruff E501 line-length limit
ruff check E501 (line too long) flagged two of the new module
docstrings. The shorter forms keep the same intent.
